+++ common/src/main/java/com/alien/common/registry/init/item/AlienBlockItems.java
@@ -14,6 +14,21 @@ public class AlienBlockItems {
         AlienResinBlocks.ABERRANT_RESIN
     );
 
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_RESIN_BRICKS = AVPBlockItems.register(
+        "aberrant_resin_bricks",
+        AlienResinBlocks.ABERRANT_RESIN_BRICKS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_RESIN_BRICK_SLAB = AVPBlockItems.register(
+        "aberrant_resin_brick_slab",
+        AlienResinBlocks.ABERRANT_RESIN_BRICK_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_RESIN_BRICK_STAIRS = AVPBlockItems.register(
+        "aberrant_resin_brick_stairs",
+        AlienResinBlocks.ABERRANT_RESIN_BRICK_STAIRS
+    );
+
     public static final AVPDeferredHolder<BlockItem> ABERRANT_RESIN_NODE = AVPBlockItems.register(
         "aberrant_resin_node",
         AlienResinBlocks.ABERRANT_RESIN_NODE
@@ -42,6 +57,21 @@ public class AlienBlockItems {
     public static final AVPDeferredHolder<BlockItem> IRRADIATED_RESIN = AVPBlockItems.register(
         "irradiated_resin",
         AlienResinBlocks.IRRADIATED_RESIN
+    );
+
+    public static final AVPDeferredHolder<BlockItem> IRRADIATED_RESIN_BRICKS = AVPBlockItems.register(
+        "irradiated_resin_bricks",
+        AlienResinBlocks.IRRADIATED_RESIN_BRICKS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> IRRADIATED_RESIN_BRICK_SLAB = AVPBlockItems.register(
+        "irradiated_resin_brick_slab",
+        AlienResinBlocks.IRRADIATED_RESIN_BRICK_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> IRRADIATED_RESIN_BRICK_STAIRS = AVPBlockItems.register(
+        "irradiated_resin_brick_stairs",
+        AlienResinBlocks.IRRADIATED_RESIN_BRICK_STAIRS
     );
 
     public static final AVPDeferredHolder<BlockItem> IRRADIATED_RESIN_NODE = AVPBlockItems.register(
@@ -73,6 +103,21 @@ public class AlienBlockItems {
         "nether_resin",
         AlienResinBlocks.NETHER_RESIN,
         new Item.Properties().fireResistant()
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_RESIN_BRICKS = AVPBlockItems.register(
+        "nether_resin_bricks",
+        AlienResinBlocks.NETHER_RESIN_BRICKS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_RESIN_BRICK_SLAB = AVPBlockItems.register(
+        "nether_resin_brick_slab",
+        AlienResinBlocks.NETHER_RESIN_BRICK_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_RESIN_BRICK_STAIRS = AVPBlockItems.register(
+        "nether_resin_brick_stairs",
+        AlienResinBlocks.NETHER_RESIN_BRICK_STAIRS
     );
 
     public static final AVPDeferredHolder<BlockItem> NETHER_RESIN_NODE = AVPBlockItems.register(
@@ -131,11 +176,56 @@ public class AlienBlockItems {
 
     public static final AVPDeferredHolder<BlockItem> RESIN_VEIN = AVPBlockItems.register("resin_vein", AlienResinBlocks.RESIN_VEIN);
 
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_RESIN_VENT = AVPBlockItems.register(
+        "aberrant_resin_vent",
+        AlienResinBlocks.ABERRANT_RESIN_VENT
+    );
+
+    public static final AVPDeferredHolder<BlockItem> IRRADIATED_RESIN_VENT = AVPBlockItems.register(
+        "irradiated_resin_vent",
+        AlienResinBlocks.IRRADIATED_RESIN_VENT
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_RESIN_VENT = AVPBlockItems.register(
+        "nether_resin_vent",
+        AlienResinBlocks.NETHER_RESIN_VENT
+    );
+
     public static final AVPDeferredHolder<BlockItem> RESIN_VENT = AVPBlockItems.register("resin_vent", AlienResinBlocks.RESIN_VENT);
 
     public static final AVPDeferredHolder<BlockItem> RESIN_WEB = AVPBlockItems.register("resin_web", AlienResinBlocks.RESIN_WEB);
 
+    public static final AVPDeferredHolder<BlockItem> RIBBED_ABERRANT_RESIN = AVPBlockItems.register(
+        "ribbed_aberrant_resin",
+        AlienResinBlocks.RIBBED_ABERRANT_RESIN
+    );
+
+    public static final AVPDeferredHolder<BlockItem> RIBBED_IRRADIATED_RESIN = AVPBlockItems.register(
+        "ribbed_irradiated_resin",
+        AlienResinBlocks.RIBBED_IRRADIATED_RESIN
+    );
+
+    public static final AVPDeferredHolder<BlockItem> RIBBED_NETHER_RESIN = AVPBlockItems.register(
+        "ribbed_nether_resin",
+        AlienResinBlocks.RIBBED_NETHER_RESIN
+    );
+
     public static final AVPDeferredHolder<BlockItem> RIBBED_RESIN = AVPBlockItems.register("ribbed_resin", AlienResinBlocks.RIBBED_RESIN);
+
+    public static final AVPDeferredHolder<BlockItem> SMOOTH_ABERRANT_RESIN = AVPBlockItems.register(
+        "smooth_aberrant_resin",
+        AlienResinBlocks.SMOOTH_ABERRANT_RESIN
+    );
+
+    public static final AVPDeferredHolder<BlockItem> SMOOTH_IRRADIATED_RESIN = AVPBlockItems.register(
+        "smooth_irradiated_resin",
+        AlienResinBlocks.SMOOTH_IRRADIATED_RESIN
+    );
+
+    public static final AVPDeferredHolder<BlockItem> SMOOTH_NETHER_RESIN = AVPBlockItems.register(
+        "smooth_nether_resin",
+        AlienResinBlocks.SMOOTH_NETHER_RESIN
+    );
 
     public static final AVPDeferredHolder<BlockItem> SMOOTH_RESIN = AVPBlockItems.register("smooth_resin", AlienResinBlocks.SMOOTH_RESIN);